feat: SecurityFilterChain 설정 변경 및 PasswordEncoder 추가 #5
