--- src/main/java/com/efub/dhs/global/config/AuthenticationConfig.java
@@ -1,25 +1,0 @@
-package com.efub.dhs.global.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.web.SecurityFilterChain;
-
-import lombok.RequiredArgsConstructor;
-
-@Configuration
-@RequiredArgsConstructor
-@EnableWebSecurity
-public class AuthenticationConfig {
-
-	@Bean
-	public SecurityFilterChain securityFilterChain(HttpSecurity httpSecurity) throws Exception {
-		return httpSecurity
-			.authorizeRequests()
-			.anyRequest().permitAll().and()
-			.httpBasic().and()
-			.build();
-	}
-
-}
